requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
PyPDF2>=3.0.1
langchain-groq>=0.1.0
groq>=0.4.0
requests>=2.31.0
numpy>=1.24.0
